+++ backend/src/main/java/com/yata/backend/domain/yata/entity/YataStatus.java
@@ -1,0 +1,27 @@
+package com.yata.backend.domain.yata.entity;
+
+import lombok.Getter;
+
+public enum YataStatus {
+        YATA_NEOTA("너타"),
+        YATA_NATA("나타");
+
+        @Getter
+        private String status;
+
+        YataStatus(String status) {
+            this.status = status;
+        }
+        public YataStatus of(String yataStr){
+            if(status ==null) {
+                throw new IllegalArgumentException();
+            }
+            for(YataStatus ys : YataStatus.values()){
+                if (ys.status.equals(yataStr)) {
+                    return ys;
+                }
+            }
+            throw new IllegalArgumentException("일치하는 게시글 상태가 없습니다.");
+        }
+    }
+
