+++ backend/src/main/java/com/yata/backend/global/exception/ExceptionCode.java
@@ -24,7 +24,8 @@ public enum ExceptionCode {
    POST_NONE(404,"POST_NONE"),
    YATA_NONE(404,"YATA_NONE"),
    YATA_IS_NOT_MODIFIABLE_STATUS(400,"YATA_IS_NOT_MODIFIABLE_STATUS"),
-   PAYMENT_NOT_ENOUGH_POINT(400,"PAYMENT_NOT_ENOUGH_POINT");
+   PAYMENT_NOT_ENOUGH_POINT(400,"PAYMENT_NOT_ENOUGH_POINT"),
+   YATA_STATUS_NONE(400,"YATA_STATUS_NONE");
 
    @Getter
    private final int code;
